menu: split files
Split out some items.  The main goal is testability of the json parts.
